feat: 구현 상세 추상화
